requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
fsspec==2025.9.0
google-auth==2.40.3
google-genai==1.38.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
llvmlite==0.45.0
MarkupSafe==3.0.2
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.5
numba==0.62.0
numpy==2.3.3
openai-whisper==20250625
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
torch==2.8.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yt-dlp==2025.9.5
